vhost: fix server mode timeout mechanism
Before this commit, RPC vf add in the server mode can cause segfault and
server mode timeout is not working. This commit fixes the issue.

RM: 3875481
